requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.29.0
pandas
numpy
scipy
openai
python-dotenv
